UBUNTU: [Packaging] neuter the signing tarball
We now produce the signing tarball from the generate package so we no
longer this to be produced from the main package.  However, in the short
term we wish to produce and publish out empty signing tarballs.  We will
do this until we have cleared out the dists versions.  Later we will
eliminate the tarball from this package entirely.

BugLink: https://bugs.launchpad.net/bugs/2012776
Signed-off-by: Andy Whitcroft <apw@canonical.com>