UBUNTU: SAUCE: apparmor4.0.0 [20/53]: 6.15 apparmor-next: apparmor: add additional flags to extended permission.
BugLink: http://bugs.launchpad.net/bugs/2028253

This is a step towards merging the file and policy state machines.

With the switch to extended permissions the state machine's ACCEPT2
table became unused freeing it up to store state specific flags. The
first flags to be stored are FLAG_OWNER and FLAG other which paves the
way towards merging the file and policydb perms into a single
permission table.

Currently Lookups based on the objects ownership conditional will
still need separate fns, this will be address in a following patch.

Signed-off-by: John Johansen <john.johansen@canonical.com>