UBUNTU: SAUCE: apparmor4.0.0 [63/84]: apparmor: ensure labels with more than one entry have correct flags
BugLink: http://bugs.launchpad.net/bugs/2028253

labels containing more than one entry need to accumulate flag info
from profiles that the label is constructed from. This is done
correctly for labels created by a merge but is not being done for
labels created by an update or directly created via a parse.

This technically is a bug fix, however the effect in current code is
to cause early unconfined bail out to not happen (ie. its slower) on
labels that were created via update or a parse.

Signed-off-by: John Johansen <john.johansen@canonical.com>