io_uring: remove unconditional looping in local task_work handling
BugLink: https://bugs.launchpad.net/bugs/2060097

[ Upstream commit 9fe3eaea4a3530ca34a8d8ff00b1848c528789ca ]

If we have a ton of notifications coming in, we can be looping in here
for a long time. This can be problematic for various reasons, mostly
because we can starve userspace. If the application is waiting on N
events, then only re-run if we need more events.

Fixes: c0e0d6ba25f1 ("io_uring: add IORING_SETUP_DEFER_TASKRUN")
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit a352d5a59f4f8812d83bfc50d48bd3a517be1791)
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>